# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "saitokdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/rate/add": {
            "post": {
                "tags": [
                    "Rate"
                ],
                "summary": "Add a rate",
                "operationId": "8ad80d422e98ef1e13c9a80dd6786a0e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rate": {
                                        "type": "string"
                                    },
                                    "comment": {
                                        "type": "string"
                                    },
                                    "buyer": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "job": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "doer": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/rate/byBuyerId/{userId}/{page}": {
            "get": {
                "tags": [
                    "Rate"
                ],
                "summary": "Get user rates",
                "operationId": "f9d0723a67f4c0d810f673d2a4eef363",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/rate/byDoerId/{userId}/{page}": {
            "get": {
                "tags": [
                    "Rate"
                ],
                "summary": "Get doer rates",
                "operationId": "d92786c5bfd0a27bf89403e700509a77",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/userId"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/wishlist/update": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Update wishlist",
                "operationId": "82b697ae38ce1d6e523a3f81b3c76829",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filter": {
                                        "properties": {
                                            "fromPrice": {
                                                "type": "string"
                                            },
                                            "toPrice": {
                                                "type": "string"
                                            },
                                            "fromDistance": {
                                                "type": "string"
                                            },
                                            "toDistance": {
                                                "type": "string"
                                            },
                                            "categories": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "currentLocation": {
                                        "properties": {
                                            "lat": {
                                                "type": "string"
                                            },
                                            "lng": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/wishlist/get": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get wishlists",
                "operationId": "ba9bb1b6dcee93b80a830b4d6c6df9d1",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
...
            "page": {
                "name": "page",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
            "userId": {
                "name": "userId",
                "in": "path",
                "allowReserved": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}